Change pre- and post- log messages
This is done in preparation for being able to pop up the log buffer with
the first new line appearing at the top of the window.

Don't have newlines at the beginning of the pre-message.

Finish the post-message with a newline, so that line will be the first of
the new output.

Do this because we can't change buffer in the current window until after
output has been generated (because of the need to retain buffer-local
variables such as `cider-ns-refresh-before-fn`.